No changes made in this commit.
The commit log message for two commits ago incorrectly states that new cross-file functionality was added. That commit actually fixed a shortcoming in the cross-file check. It was using a line-by-line scanning strategy wherein the regex was applied to a single line of text at a time which is undesirable in cases where you want to use a DOIALL regular expression.

I've added an option when you define cross-file custom rules to allow you to override the default single-line-at-a-time scan behavior with a whole-file scan strategy. Note that the whole-file scan strategy is not entirely new. It already existed in some of this plugin's other "Check" classes.